requirements.txt
flask
torch
torchvision
open_clip
scikit-learn
Pillow
